backend/requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-multipart==0.0.6

# Async Support
aiohttp==3.9.1
aiofiles==23.2.1

# WORKING LangChain combination
langchain==0.1.16
langchain-community==0.0.34
langchain-core==0.1.46
langsmith==0.1.48

# Embeddings and LLM
sentence-transformers==2.2.2
transformers==4.36.0
torch==2.1.0
tokenizers==0.15.0

# Vector Database
chromadb==0.4.18

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Redis
redis==5.0.1

# Data Processing
numpy==1.24.3
scikit-learn==1.3.2
pandas==2.1.4

# API utilities
httpx==0.25.2
requests==2.31.0

# Additional LangChain dependencies
dataclasses-json==0.6.3
marshmallow==3.20.1
typing-inspect==0.9.0
typing-extensions==4.9.0
jsonpatch==1.33
tenacity==8.2.3

# Utilities
pyyaml==6.0.1
python-json-logger==2.0.7
joblib==1.3.2